addresses requested changes and organizes post sections
